refactor(chat): 파일 이동 및 사용하지 않는 import문 제거
- ChatApiResponse : entity폴더에서 dto폴더로 이동
- ChatRoomRequest : 사용하지 않는 import문 제거

#4 #6
