+++ web-ide/src/main/java/goorm/webide/chat/dto/ChatRoomRequest.java
@@ -1,7 +1,6 @@
 package goorm.webide.chat.dto;
 
 import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.NotEmpty;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
